requirements.txt
pyyaml
nibabel
matplotlib
scipy
numpy
scikit-image
dipy
medpy